+++ src/modules/CommentContainer/CommentCommon/index.ts
@@ -1,0 +1,7 @@
+export * from './Content.client';
+export * from './DisplayLike.client';
+export * from './Header.client';
+export * from './LikeIcon.client';
+export * from './ReplyShowButton.client';
+export * from './ReplySubmitButton.client';
+export * from './UserProfile.client';
