+++ src/modules/CommentContainer/CommentReplyContainer/index.ts
@@ -1,0 +1,2 @@
+export * from './CommentReply.client';
+export * from './CommentReplyList.client';
